% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coef.spmixW}
\alias{coef.spmixW}
\title{Coefficient Extractor for spmixW Objects}
\usage{
\method{coef}{spmixW}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"spmixW"}.}

\item{...}{Further arguments (ignored).}
}
\value{
Named numeric vector of posterior means.
}
\description{
Coefficient Extractor for spmixW Objects
}
