% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_bma.R
\name{coef.spmixW_bma}
\alias{coef.spmixW_bma}
\title{Coefficient Extractor for spmixW_bma Objects}
\usage{
\method{coef}{spmixW_bma}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"spmixW_bma"}.}

\item{...}{Further arguments (ignored).}
}
\value{
Named numeric vector of BMA-averaged posterior means.
}
\description{
Coefficient Extractor for spmixW_bma Objects
}
