% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom_methods.R
\name{glance.spmixW}
\alias{glance.spmixW}
\title{Glance Method for spmixW Objects}
\usage{
glance.spmixW(x, ...)
}
\arguments{
\item{x}{An object of class \code{"spmixW"}.}

\item{...}{Additional arguments (ignored).}
}
\value{
A one-row data frame.
}
\description{
Returns a one-row data frame of model-level summary statistics.
}
