% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom_methods.R
\name{glance.spmixW_bma}
\alias{glance.spmixW_bma}
\title{Glance Method for spmixW_bma Objects}
\usage{
glance.spmixW_bma(x, ...)
}
\arguments{
\item{x}{An object of class \code{"spmixW_bma"}.}

\item{...}{Additional arguments (ignored).}
}
\value{
A one-row data frame.
}
\description{
Glance Method for spmixW_bma Objects
}
