% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{logLik.spmixW}
\alias{logLik.spmixW}
\title{Log-Likelihood for spmixW Objects}
\usage{
\method{logLik}{spmixW}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"spmixW"}.}

\item{...}{Further arguments (ignored).}
}
\value{
The log-likelihood evaluated at the posterior mean.
}
\description{
Log-Likelihood for spmixW Objects
}
