% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.spmixW}
\alias{plot.spmixW}
\title{Plot Method for spmixW Objects}
\usage{
\method{plot}{spmixW}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"spmixW"}.}

\item{...}{Further arguments (ignored).}
}
\value{
Invisible NULL.
}
\description{
Produces a 2x2 diagnostic plot layout: trace and density for the spatial
parameter, trace for sigma^2, and an effects comparison (or gamma densities
for convex combination models).
}
