% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_bma.R
\name{plot.spmixW_bma}
\alias{plot.spmixW_bma}
\title{Plot Method for spmixW_bma Objects}
\usage{
\method{plot}{spmixW_bma}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"spmixW_bma"}.}

\item{...}{Further arguments (ignored).}
}
\value{
Invisible NULL.
}
\description{
Produces a multi-panel plot showing model probabilities and BMA
posterior densities.
}
