% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.spmixW}
\alias{print.spmixW}
\title{Print Method for spmixW Objects}
\usage{
\method{print}{spmixW}(x, digits = 4, ...)
}
\arguments{
\item{x}{An object of class \code{"spmixW"}.}

\item{digits}{Integer: number of significant digits to print.}

\item{...}{Further arguments (ignored).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print Method for spmixW Objects
}
