% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_bma.R
\name{print.spmixW_bma}
\alias{print.spmixW_bma}
\title{Print Method for spmixW_bma Objects}
\usage{
\method{print}{spmixW_bma}(x, digits = 4, ...)
}
\arguments{
\item{x}{An object of class \code{"spmixW_bma"}.}

\item{digits}{Integer: significant digits.}

\item{...}{Further arguments (ignored).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print Method for spmixW_bma Objects
}
