% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bma.R
\name{sdm_conv_bma}
\alias{sdm_conv_bma}
\title{Bayesian Model Averaging for SDM Convex Combination Panel}
\usage{
sdm_conv_bma(
  y,
  X,
  Wlist,
  N,
  Time,
  ndraw = 25000L,
  nomit = 5000L,
  prior = list()
)
}
\arguments{
\item{y}{Numeric vector of length NT.}

\item{X}{Numeric matrix NT x k.}

\item{Wlist}{A list of M spatial weight matrices (each N x N).}

\item{N}{Integer: number of cross-sectional units.}

\item{Time}{Integer: number of time periods.}

\item{ndraw}{Integer: MCMC draws per model (recommend >= 10000).}

\item{nomit}{Integer: burn-in draws per model.}

\item{prior}{List of prior hyperparameters (passed to each model).}
}
\value{
An S3 object of class \code{"spmixW_bma"}.
}
\description{
Bayesian Model Averaging for SDM Convex Combination Panel
}
