% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.spmixW}
\alias{summary.spmixW}
\title{Summary Method for spmixW Objects}
\usage{
\method{summary}{spmixW}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"spmixW"}.}

\item{...}{Further arguments (ignored).}
}
\value{
Invisibly returns \code{object} with MCMC diagnostics appended.
}
\description{
Summary Method for spmixW Objects
}
