% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_bma.R
\name{summary.spmixW_bma}
\alias{summary.spmixW_bma}
\title{Summary Method for spmixW_bma Objects}
\usage{
\method{summary}{spmixW_bma}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"spmixW_bma"}.}

\item{...}{Further arguments (ignored).}
}
\value{
Invisibly returns \code{object}.
}
\description{
Summary Method for spmixW_bma Objects
}
