% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom_methods.R
\name{tidy.spmixW_bma}
\alias{tidy.spmixW_bma}
\title{Tidy Method for spmixW_bma Objects}
\usage{
tidy.spmixW_bma(x, ...)
}
\arguments{
\item{x}{An object of class \code{"spmixW_bma"}.}

\item{...}{Additional arguments (ignored).}
}
\value{
A data frame with BMA-averaged coefficients and model probabilities.
}
\description{
Tidy Method for spmixW_bma Objects
}
